add files in a window
