+++ docs/todo.md
@@ -11,10 +11,13 @@
  - [ ] put the name of the session in topbar
  - [ ] rather than system as the username, put the name of the session
  - [ ] why do finetunes stick around in GPU memory? once they're done they should exit right?
+ - [ ] add a restart button whilst doing a fine-tune so if things get stuck we can restart
+   - [ ] possibly only show this if we've not seen any progress for > 30 seconds
  - [ ] the session page scrolls to the bottom randomly
  - [ ] quite often there's a model ready to serve and a new one gets spun up on the other node - maybe the clocks are drifting between the machines so the 2 second head start doesn't work? or the python processes aren't polling every 100ms or something?
  - [ ] speedway empty error https://mlops-community.slack.com/archives/C0675EX9V2Q/p1701725991656799
  - [ ] empty response messages error https://mlops-community.slack.com/archives/C0675EX9V2Q/p1701727773319809
+ - [ ] we are getting nignx 500's in the runner "load session from api" handler https://mlops-community.slack.com/archives/C0675EX9V2Q/p1702369315736539
  - [ ] performance of auto-save before login (image fine tune text is slow)
  - [ ] for session updates check we are on the same page
    - [ ] whilst we are on one page and another session is processing - it's updating the page we are on with the wrong session
